requirements/requirements.txt
Django==4.2.7
daphne==4.1.2
psycopg>=3.1.8
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
django-cors-headers==4.3.1
django-filter
celery==5.4.0
jsonschema==4.21.1
aiohttp==3.9.3
pyjwe==1.0.0
cryptography==42.0.7

# For Amazon S3
boto3==1.34
